Fix the unit tests and the test directory
