+++ src/test/java/org/sefglobal/scholarx/service/ProgramServiceTest.java
@@ -8,15 +8,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.sefglobal.scholarx.exception.BadRequestException;
 import org.sefglobal.scholarx.exception.NoContentException;
 import org.sefglobal.scholarx.exception.ResourceNotFoundException;
-import org.sefglobal.scholarx.model.Mentee;
-import org.sefglobal.scholarx.model.Mentor;
-import org.sefglobal.scholarx.model.Profile;
-import org.sefglobal.scholarx.model.Program;
-import org.sefglobal.scholarx.repository.MenteeRepository;
-import org.sefglobal.scholarx.repository.MentorRepository;
-import org.sefglobal.scholarx.repository.ProfileRepository;
-import org.sefglobal.scholarx.repository.ProgramRepository;
-import org.sefglobal.scholarx.util.EnrolmentState;
+import org.sefglobal.scholarx.model.*;
+import org.sefglobal.scholarx.repository.*;
 import org.sefglobal.scholarx.util.ProgramState;
 
 import java.util.ArrayList;
@@ -27,6 +20,7 @@ import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.anyLong;
 import static org.mockito.Mockito.doReturn;
 
@@ -39,6 +33,8 @@ public class ProgramServiceTest {
     @Mock
     private MentorRepository mentorRepository;
     @Mock
+    private MentorResponseRepository mentorResponseRepository;
+    @Mock
     private MenteeRepository menteeRepository;
     @InjectMocks
     private ProgramService programService;
@@ -46,13 +42,11 @@ public class ProgramServiceTest {
     private final Long profileId = 1L;
     private final Program program =
             new Program("SCHOLARX-2020", "SCHOLARX program of 2020",
-                        "http://scholarx/images/SCHOLARX-2020",
-                        "http://scholarx/SCHOLARX-2020/home", ProgramState.CREATED);
-    private final Mentor mentor =
-            new Mentor("Sample application",
-                       "Sample prerequisites");
-    private final Profile profile =
-            new Profile();
+                    "https://scholarx/images/SCHOLARX-2020",
+                    "https://scholarx/SCHOLARX-2020/home", ProgramState.CREATED);
+    private final Mentor mentor = new Mentor();
+    private final Profile profile = new Profile();
+    private final List<MentorResponse> mentorResponses = new ArrayList<>();
 
     @Test
     void updateState_withValidData_thenReturnUpdatedData()
@@ -155,8 +149,8 @@ public class ProgramServiceTest {
             throws ResourceNotFoundException, BadRequestException {
         final Program program =
                 new Program("SCHOLARX-2020", "SCHOLARX program of 2020",
-                            "http://scholarx/images/SCHOLARX-2020",
-                            "http://scholarx/SCHOLARX-2020/home",
+                        "https://scholarx/images/SCHOLARX-2020",
+                        "https://scholarx/SCHOLARX-2020/home",
                             ProgramState.MENTOR_APPLICATION);
 
         doReturn(Optional.of(program))
@@ -169,7 +163,7 @@ public class ProgramServiceTest {
                 .when(mentorRepository)
                 .save(any(Mentor.class));
 
-        Mentor savedMentor = programService.applyAsMentor(programId, profileId, mentor);
+        Mentor savedMentor = programService.applyAsMentor(programId, profileId, mentorResponses);
         assertThat(savedMentor).isNotNull();
     }
 
@@ -180,7 +174,7 @@ public class ProgramServiceTest {
                 .findById(anyLong());
 
         Throwable thrown = catchThrowable(
-                () -> programService.applyAsMentor(programId, profileId, mentor));
+                () -> programService.applyAsMentor(programId, profileId, mentorResponses));
         assertThat(thrown)
                 .isInstanceOf(ResourceNotFoundException.class)
                 .hasMessage("Error, Unable to apply as a mentor. " +
@@ -194,7 +188,7 @@ public class ProgramServiceTest {
                 .findById(anyLong());
 
         Throwable thrown = catchThrowable(
-                () -> programService.applyAsMentor(programId, profileId, mentor));
+                () -> programService.applyAsMentor(programId, profileId, mentorResponses));
         assertThat(thrown)
                 .isInstanceOf(BadRequestException.class)
                 .hasMessage("Error, Unable to apply as a mentor. " +
@@ -205,8 +199,8 @@ public class ProgramServiceTest {
     void applyAsMentor_withUnavailableProfile_thenThrowResourceNotFound() {
         final Program program =
                 new Program("SCHOLARX-2020", "SCHOLARX program of 2020",
-                        "http://scholarx/images/SCHOLARX-2020",
-                        "http://scholarx/SCHOLARX-2020/home",
+                        "https://scholarx/images/SCHOLARX-2020",
+                        "https://scholarx/SCHOLARX-2020/home",
                         ProgramState.MENTOR_APPLICATION);
 
         doReturn(Optional.of(program))
@@ -217,7 +211,7 @@ public class ProgramServiceTest {
                 .findById(anyLong());
 
         Throwable thrown = catchThrowable(
-                () -> programService.applyAsMentor(programId, profileId, mentor));
+                () -> programService.applyAsMentor(programId, profileId, mentorResponses));
         assertThat(thrown)
                 .isInstanceOf(ResourceNotFoundException.class)
                 .hasMessage("Error, Unable to apply as a mentor. " +
@@ -239,49 +233,35 @@ public class ProgramServiceTest {
     }
 
     @Test
-    void updateMentorData_withValidData_thenReturnUpdatedData()
+    void editMentorResponses_withValidData_thenReturnUpdatedData()
             throws ResourceNotFoundException, BadRequestException {
-        mentor.setState(EnrolmentState.PENDING);
+        MentorResponse mentorResponse = new MentorResponse();
+        mentorResponse.setId(new MentorResponseId());
         program.setState(ProgramState.MENTOR_APPLICATION);
         mentor.setProgram(program);
         doReturn(Optional.of(mentor))
                 .when(mentorRepository)
                 .findByProfileIdAndProgramId(anyLong(), anyLong());
-        doReturn(mentor)
-                .when(mentorRepository)
-                .save(any(Mentor.class));
+        doReturn(mentorResponses)
+                .when(mentorResponseRepository)
+                .saveAll(anyList());
 
-        Mentor savedMentor = programService.updateMentorData(profileId, programId, mentor);
-        assertThat(savedMentor).isNotNull();
+        List<MentorResponse> savedMentorResponses = programService.editMentorResponses(profileId, programId, mentorResponses);
+        assertThat(savedMentorResponses).isNotNull();
     }
 
     @Test
-    void updateMentorData_withUnavailableData_thenThrowResourceNotFound() {
+    void editMentorResponses_withUnavailableData_thenThrowResourceNotFound() {
         doReturn(Optional.empty())
                 .when(mentorRepository)
                 .findByProfileIdAndProgramId(anyLong(), anyLong());
 
         Throwable thrown = catchThrowable(
-                () -> programService.updateMentorData(profileId, programId, mentor));
+                () -> programService.editMentorResponses(profileId, programId, mentorResponses));
         assertThat(thrown)
                 .isInstanceOf(ResourceNotFoundException.class)
-                .hasMessage("Error, Mentor by profile id: 1 and program id: 1 cannot be updated. " +
+                .hasMessage("Error, Mentor by profile id: 1 and program id: 1 cannot be found. " +
                             "Mentor doesn't exist.");
-    }
-
-    @Test
-    void updateMentorData_withUnsuitableData_thenThrowBadRequest() {
-        mentor.setState(EnrolmentState.APPROVED);
-        doReturn(Optional.of(mentor))
-                .when(mentorRepository)
-                .findByProfileIdAndProgramId(anyLong(), anyLong());
-
-        Throwable thrown = catchThrowable(
-                () -> programService.updateMentorData(profileId, programId, mentor));
-        assertThat(thrown)
-                .isInstanceOf(BadRequestException.class)
-                .hasMessage("Error, Application cannot be updated. " +
-                            "Mentor is not in a valid state.");
     }
 
     @Test
